% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_rebates.R
\docType{data}
\name{data.rebate_on_investments}
\alias{data.rebate_on_investments}
\title{tax rebates under step 2}
\format{
data.rebate_on_investments: A data frame:
\describe{
  \item{rebate.id}{rebate identification number}
  \item{rebate.code}{rebate code/short name}
  \item{rebate.limit}{maximum rebate}
  \item{tax_regime}{old or new regime}
  \item{financial_year}{financial year}
  \item{rebate.info}{comments or additional information}
}
}
\usage{
data.rebate_on_investments
}
\description{
tax rebates after tax slab adjustments
}
\keyword{datasets}
